+++ src/main/java/fr/diginamic/appliweb/entities/Departement.java
@@ -31,7 +31,7 @@ public class Departement {
 
     // GETTERS / SETTERS
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
